TA: CB fix signed ta installation on t5w platform [1/1]
PD#TV-77838

Problem:
signed ta not installed on t5w platform

Solution:
implements CHIPSET_NAME for t5d/t5w
implements TDK_VERSION for t5w
fix compile issue of aucpu-fw/media-modules for t5d

Verify:
t5d/t5w

+++ recipes-kernel/media_modules/media-modules_5.4.bb
@@ -23,7 +23,11 @@ do_install() {
     unset CFLAGS CPPFLAGS CXXFLAGS LDFLAGS
     mkdir -p ${MEDIADIR} ${FIRMWAREDIR}
     find ${S}/drivers/ -name *.ko | xargs -i install -m 0666 {} ${MEDIADIR}
-    install -m 0666 ${MEDIA_MODULES_UCODE_BIN} ${FIRMWAREDIR}
+    if [ -f ${MEDIA_MODULES_UCODE_BIN} ]; then
+        install -m 0666 ${MEDIA_MODULES_UCODE_BIN} ${FIRMWAREDIR}
+    else
+        install -m 0666 ${S}/firmware/video_ucode.bin ${FIRMWAREDIR}
+    fi
 }
 
 FILES:${PN} = " \
